plugins: Remove legacy slideshow source
This removes the old slideshow source. Since the properties are
the same for both of the new and old slideshows, they are
backwards compatible.
