+++ plugins/image-source/CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 3.28...3.30)
 add_library(image-source MODULE)
 add_library(OBS::image-source ALIAS image-source)
 
-target_sources(image-source PRIVATE color-source.c image-source.c obs-slideshow.c obs-slideshow-mk2.c)
+target_sources(image-source PRIVATE color-source.c image-source.c obs-slideshow-mk2.c)
 
 target_link_libraries(image-source PRIVATE OBS::libobs $<$<PLATFORM_ID:Windows>:OBS::w32-pthreads>)
