+++ plugins/image-source/obs-slideshow-mk2.c
@@ -234,7 +234,17 @@ static inline void add_file(image_file_array_t *new_files, const char *path)
 	da_push_back(*new_files, &data);
 }
 
-extern bool valid_extension(const char *ext);
+bool valid_extension(const char *ext)
+{
+	if (!ext)
+		return false;
+	return astrcmpi(ext, ".bmp") == 0 || astrcmpi(ext, ".tga") == 0 || astrcmpi(ext, ".png") == 0 ||
+	       astrcmpi(ext, ".jpeg") == 0 || astrcmpi(ext, ".jpg") == 0 ||
+#ifdef _WIN32
+	       astrcmpi(ext, ".jxr") == 0 ||
+#endif
+	       astrcmpi(ext, ".gif") == 0;
+}
 
 /* transition to new source. assumes cur has already been set to new file */
 static void do_transition(void *data, bool to_null)
@@ -1143,6 +1153,36 @@ static enum gs_color_space ss_video_get_color_space(void *data, size_t count,
 
 	return space;
 }
+
+struct obs_source_info slideshow_info = {
+	.id = "slideshow",
+	.type = OBS_SOURCE_TYPE_INPUT,
+	.output_flags = OBS_SOURCE_VIDEO | OBS_SOURCE_CUSTOM_DRAW | OBS_SOURCE_COMPOSITE |
+			OBS_SOURCE_CONTROLLABLE_MEDIA | OBS_SOURCE_CAP_OBSOLETE,
+	.get_name = ss_getname,
+	.create = ss_create,
+	.destroy = ss_destroy,
+	.update = ss_update,
+	.activate = ss_activate,
+	.deactivate = ss_deactivate,
+	.video_render = ss_video_render,
+	.video_tick = ss_video_tick,
+	.audio_render = ss_audio_render,
+	.enum_active_sources = ss_enum_sources,
+	.get_width = ss_width,
+	.get_height = ss_height,
+	.get_defaults = ss_defaults,
+	.get_properties = ss_properties,
+	.missing_files = ss_missingfiles,
+	.icon_type = OBS_ICON_TYPE_SLIDESHOW,
+	.media_play_pause = ss_play_pause,
+	.media_restart = ss_restart,
+	.media_stop = ss_stop,
+	.media_next = ss_next_slide,
+	.media_previous = ss_previous_slide,
+	.media_get_state = ss_get_state,
+	.video_get_color_space = ss_video_get_color_space,
+};
 
 struct obs_source_info slideshow_info_mk2 = {
 	.id = "slideshow",
